+++ backend/__tests__/acceptance/io.spec.cjs
@@ -9,7 +9,7 @@
 const { mockRequest } = require('@gardener-dashboard/request')
 const { Store } = require('@gardener-dashboard/kube-client')
 const { mockListIssues, mockListComments } = require('@octokit/core')
-const { pEvent } = require('../../p-event-helper')
+const { pEvent } = require('../../__test_helpers__/p-event')
 const createError = require('http-errors')
 const tickets = require('../../dist/lib/services/tickets')
 const cache = require('../../dist/lib/cache')
